+++ log-processor/src/test/java/core/log/service/IndexServiceTest.java
@@ -1,0 +1,35 @@
+package core.log.service;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.time.LocalDate;
+import java.time.Month;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * @author neo
+ */
+class IndexServiceTest {
+    private IndexService indexService;
+
+    @BeforeEach
+    void createIndexService() {
+        indexService = new IndexService();
+    }
+
+    @Test
+    void indexName() {
+        assertThat(indexService.indexName("action", LocalDate.of(2016, Month.JANUARY, 15))).isEqualTo("action-2016.01.15");
+        assertThat(indexService.indexName("trace", LocalDate.of(2018, Month.AUGUST, 1))).isEqualTo("trace-2018.08.01");
+    }
+
+    @Test
+    void createdDate() {
+        assertThat(indexService.createdDate("action-2016.02.03")).get().isEqualTo(LocalDate.of(2016, Month.FEBRUARY, 3));
+        assertThat(indexService.createdDate("stat-2015.11.15")).get().isEqualTo(LocalDate.of(2015, Month.NOVEMBER, 15));
+        assertThat(indexService.createdDate("metricbeat-6.3.2-2018.08.19")).get().isEqualTo(LocalDate.of(2018, Month.AUGUST, 19));
+        assertThat(indexService.createdDate(".kibana")).isNotPresent();
+    }
+}