+++ log-processor/src/test/java/core/log/service/StatServiceTest.java
@@ -15,7 +15,7 @@ import java.time.Month;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author neo
@@ -23,7 +23,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 class StatServiceTest extends IntegrationTest {
     @Inject
     StatService statService;
-
+    @Inject
+    IndexService indexService;
     @Inject
     ElasticSearchType<StatDocument> statType;
 
@@ -31,11 +32,11 @@ class StatServiceTest extends IntegrationTest {
     void index() {
         StatMessage message = message("1");
 
-        LocalDate now = LocalDate.of(2017, Month.OCTOBER, 10);
+        var now = LocalDate.of(2017, Month.OCTOBER, 10);
         statService.index(List.of(message), now);
 
         StatDocument stat = statDocument(now, message.id);
-        assertEquals(message.stats, stat.stats);
+        assertThat(stat.stats).isEqualTo(message.stats);
     }
 
     @Test
@@ -45,22 +46,22 @@ class StatServiceTest extends IntegrationTest {
             messages.add(message("bulk-" + i));
         }
 
-        LocalDate now = LocalDate.of(2017, Month.OCTOBER, 10);
+        var now = LocalDate.of(2017, Month.OCTOBER, 10);
         statService.index(messages, now);
 
         StatDocument stat = statDocument(now, messages.get(0).id);
-        assertEquals(messages.get(0).stats, stat.stats);
+        assertThat(stat.stats).isEqualTo(messages.get(0).stats);
     }
 
     private StatDocument statDocument(LocalDate now, String id) {
-        GetRequest request = new GetRequest();
-        request.index = IndexName.name("stat", now);
+        var request = new GetRequest();
+        request.index = indexService.indexName("stat", now);
         request.id = id;
         return statType.get(request).orElseThrow(() -> new Error("not found"));
     }
 
     private StatMessage message(String id) {
-        StatMessage message = new StatMessage();
+        var message = new StatMessage();
         message.id = id;
         message.date = Instant.now();
         message.stats = Map.of("thread_count", 10d);